+++ tests/core/tests/test_fields.py
@@ -110,6 +110,14 @@ class FieldTest(TestCase):
         method_name = field.get_dehydrate_method(resource_field_name)
         self.assertEqual(method_name, custom_dehydrate_method)
 
+    def test_get_dehydrate_method_with_callable(self):
+        field = fields.Field(
+            attribute="foo", column_name="bar", dehydrate_method=lambda x: x
+        )
+        resource_field_name = "field"
+        method = field.get_dehydrate_method(resource_field_name)
+        self.assertTrue(callable(method))
+
     def testget_dehydrate_method_without_params_raises_attribute_error(self):
         field = fields.Field(attribute="foo", column_name="bar")